requirements.txt
cffi==1.11.5
gevent==1.3.2.post0
greenlet==0.4.13
pycparser==2.18
six==1.11.0
websocket-client==0.48.0
